pyproject.toml
[tool.black]
line-length = 119
target-version = ['py39']
include = '\.pyi?$'
diff=false
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 119
max-complexity = 18
select = "B,C,E,F,W,T4,B9"
ignore = "E731"

[tool.isort]
line_length = 119
multi_line_output = 3
include_trailing_comma = true
known_third_party = "fastapi,tortoise,ulid,uvicorn"


[tool.poetry]
name = "svc"
version = "0.1.0"
description = ""
authors = ["Pritam Soni <23050213+pritamsoni-hsr@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "svc"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.86.0"
uvicorn = "^0.19.0"
tortoise = "^0.1.1"
google-auth = "^2.14.1"
requests = "^2.28.1"
pyjwt = {extras = ["cryptography"], version = "^2.6.0"}


[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
mypy = "^0.990"
pytest = "^7.2.0"
flake8 = "^5.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
